Made significant modifications to algorithm's op.
Made some significatn changes to how my algorithm operates
I realized that before, my generator was cycling throuhg combinations
of ALL the filters regardless of which ones were set when starting.
I addressed this buy creating bins for the filters and their options.
The bins are filled by the generator, so that the SQL queries can be
made by picking up the entire bin.